Make a module specifically for Modbus RTU stuff.
