+++ haskell-modbus.cabal
@@ -12,7 +12,7 @@ build-type:          Simple
 cabal-version:       >=1.10
 
 library
-  exposed-modules:     Data.Modbus, Data.Digest.CRC16
+  exposed-modules:     Data.Modbus, Data.ModbusRtu, Data.Digest.CRC16
   default-language:    Haskell2010
   ghc-options: -Wall
   build-depends:       base >= 4 && < 5,
